Очистка репозитория от ненужных файлов
