--- bacs/endgame.txt
@@ -1,31 +1,0 @@
-str=2 end=8 clr=ff87b4 mutstat=93 mutbeh=100 behaviour={ 23 25 10 17 3 7 24 2 24 21 12 22 0 13 25 12 17 9 25 22 }
-str=7 end=12 clr=ff757a mutstat=85 mutbeh=100 behaviour={ 6 13 18 0 23 9 23 25 11 21 5 25 0 21 17 22 12 9 8 16 }
-str=9 end=9 clr=ff7676 mutstat=83 mutbeh=100 behaviour={ 6 10 3 18 23 11 23 25 11 21 25 25 0 21 17 22 12 9 8 19 }
-str=4 end=5 clr=ff7cee mutstat=83 mutbeh=100 behaviour={ 6 2 2 4 8 22 23 25 18 21 17 25 0 21 17 22 12 24 12 25 }
-str=2 end=-2 clr=FFFFFF mutstat=124 mutbeh=100 behaviour={ 3 25 17 2 22 14 21 25 25 21 7 25 23 15 2 6 9 21 11 0 }
-str=15 end=9 clr=ff737a mutstat=85 mutbeh=100 behaviour={ 6 7 6 18 23 13 23 25 12 21 3 25 0 21 13 22 0 9 11 6 }
-str=5 end=8 clr=FFFFFF mutstat=80 mutbeh=100 behaviour={ 6 6 2 4 8 22 23 25 11 21 17 25 0 21 17 22 12 24 12 25 }
-str=1 end=14 clr=ff5de4 mutstat=92 mutbeh=100 behaviour={ 23 25 10 23 17 13 3 25 23 21 12 25 0 22 21 22 12 9 19 23 }
-str=7 end=8 clr=ff78ce mutstat=88 mutbeh=100 behaviour={ 6 0 8 24 10 22 23 25 6 21 9 25 0 21 17 22 12 9 16 23 }
-str=1 end=8 clr=ff6c44 mutstat=93 mutbeh=100 behaviour={ 23 9 23 19 23 11 6 2 5 21 14 25 21 21 0 22 19 19 13 23 }
-str=8 end=12 clr=ff73b6 mutstat=90 mutbeh=100 behaviour={ 6 0 13 11 23 9 23 25 11 21 5 25 0 21 17 22 12 9 16 16 }
-str=1 end=8 clr=ff73e0 mutstat=96 mutbeh=100 behaviour={ 23 5 21 23 19 25 14 21 17 21 15 25 2 21 0 22 3 9 17 23 }
-str=8 end=11 clr=ff7aa5 mutstat=84 mutbeh=100 behaviour={ 6 21 20 14 21 13 23 25 11 21 5 25 0 21 5 22 12 3 16 3 }
-str=2 end=15 clr=ff5618 mutstat=88 mutbeh=100 behaviour={ 23 25 0 23 17 13 9 25 23 21 12 25 0 19 25 22 12 9 5 15 }
-str=4 end=4 clr=ff7be9 mutstat=102 mutbeh=100 behaviour={ 23 5 22 23 21 25 14 21 21 21 19 25 1 21 0 22 3 6 10 2 }
-str=1 end=11 clr=ff582c mutstat=94 mutbeh=100 behaviour={ 23 25 10 23 9 21 11 25 21 21 12 25 0 22 21 22 12 21 5 15 }
-str=5 end=12 clr=ff7666 mutstat=84 mutbeh=100 behaviour={ 6 0 18 24 8 13 23 25 11 21 5 25 0 21 17 22 12 9 16 16 }
-str=3 end=15 clr=ff6537 mutstat=82 mutbeh=100 behaviour={ 23 4 9 11 9 0 13 21 7 21 1 25 18 18 9 22 19 6 23 25 }
-str=5 end=9 clr=ff77fd mutstat=80 mutbeh=100 behaviour={ 6 10 3 14 25 11 23 25 11 21 3 25 0 21 17 22 12 9 8 5 }
-str=1 end=7 clr=ff7a58 mutstat=82 mutbeh=100 behaviour={ 6 12 13 10 2 9 23 25 11 21 11 25 0 21 9 22 12 9 8 7 }
-str=3 end=4 clr=ff804e mutstat=97 mutbeh=100 behaviour={ 9 11 23 14 0 13 1 15 18 14 20 25 3 14 21 22 21 20 23 1 }
-str=2 end=10 clr=ff78c6 mutstat=108 mutbeh=100 behaviour={ 23 5 15 18 6 25 14 21 21 2 1 25 6 21 0 22 14 22 2 23 }
-str=3 end=13 clr=ff63f7 mutstat=87 mutbeh=100 behaviour={ 23 4 9 11 9 0 13 21 7 21 1 25 18 18 9 22 19 6 23 25 }
-str=0 end=0 clr=000000 mutstat=0 mutbeh=0 behaviour={ 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 }
-str=1 end=12 clr=ff784c mutstat=99 mutbeh=100 behaviour={ 23 5 15 23 25 25 14 21 17 21 15 25 1 21 0 22 3 9 13 19 }
-str=12 end=9 clr=ff73dd mutstat=88 mutbeh=100 behaviour={ 6 21 1 14 21 13 23 25 11 21 25 25 0 21 17 22 12 9 16 16 }
-str=1 end=-3 clr=FFFFFF mutstat=126 mutbeh=100 behaviour={ 3 25 17 2 24 14 14 25 25 21 7 0 23 15 2 21 9 12 11 0 }
-str=2 end=4 clr=ff7ace mutstat=83 mutbeh=100 behaviour={ 6 7 10 5 13 0 23 25 11 21 17 25 0 21 9 22 12 9 5 6 }
-str=9 end=10 clr=ff769a mutstat=84 mutbeh=100 behaviour={ 6 10 3 18 3 22 23 25 11 21 3 25 0 21 17 22 12 9 8 14 }
-str=1 end=4 clr=ff7e7f mutstat=80 mutbeh=100 behaviour={ 6 23 25 18 16 22 23 25 11 21 17 25 0 21 5 22 12 24 12 25 }
-time: 1646.5808
