+++ build.gradle
@@ -7,7 +7,7 @@ version = '1.12'
 
 apply plugin: "java"
 apply plugin: "application"
-mainClassName = "base.Bacs"
+mainClassName = "Bacs"
 
 tasks.withType(JavaCompile) {
     sourceCompatibility = JavaVersion.VERSION_1_8
