+++ src/main/java/base/BattleField.java
@@ -3,7 +3,7 @@ package base;
 import java.util.HashMap;
 
 /**
- *  Игровое поле.
+ * Игровое поле.
  * Created by valera on 08.03.17.
  */
 public class BattleField {
@@ -17,7 +17,7 @@ public class BattleField {
     private final boolean useLight;
 
     private final HashMap<Integer, Double> lightTable = new HashMap<>();
-    
+
     public BattleField(int dimension, boolean useLight) {
         this.dimension = dimension;
         cells = new BacUnit[dimension * dimension];
@@ -41,7 +41,7 @@ public class BattleField {
         }
 
         for (int i = 0; i < centerRow; i++)
-            for (int j = 0; j < centerColumn; j ++)
+            for (int j = 0; j < centerColumn; j++)
                 lightTable.put(toNum(j, i), calculateLight(j, i, centerColumn, centerRow));
 
         int initial = toNum(centerColumn, centerRow);
@@ -63,7 +63,7 @@ public class BattleField {
     }
 
     public double getLight(int x, int y) {
-        return lightTable.get(toNum(x - dimension / 2, y - dimension/2));
+        return lightTable.get(toNum(x - dimension / 2, y - dimension / 2));
     }
 
     private double calculateLight(int x, int y, int xc, int yc) {
@@ -71,7 +71,7 @@ public class BattleField {
             return 1;
         int dx = xc - x;
         int dy = yc - y;
-        double delta = Math.sqrt(dx*dx + dy*dy);
+        double delta = Math.sqrt(dx * dx + dy * dy);
         double maxDistance = Math.sqrt(2) * dimension / 2;
         return (maxDistance - delta) / maxDistance;
     }
